lib/http: split http/types.h, use enum http_version for http_write_response
